refactor: more nominal->noun
