+++ harper-core/src/linting/proper_noun_capitalization_linters.rs
