+++ harper-core/tests/run_tests.rs
@@ -40,7 +40,7 @@ create_test!(issue_109_ext.md, 0);
 create_test!(chinese_lorem_ipsum.md, 2);
 create_test!(obsidian_links.md, 2);
 create_test!(issue_267.md, 0);
-create_test!(proper_nominal_capitalization.md, 2);
+create_test!(proper_noun_capitalization.md, 2);
 create_test!(amazon_hostname.md, 0);
 create_test!(issue_159.md, 1);
 create_test!(issue_358.md, 0);
